Moved FileInPath class to FWCore/Utilities
